+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - OpenAPI users endpoint now returns the ID of the InvoicingProfile
 - Fix a bug: URL validation regex was wrong
 - Fix a bug: privileged users cannot order free carts for themselves in the store
+- Fix a bug: unable to select a new machine for an existing category
 - Fix a bug: wrong counting of minutes used when using a prepaid pack
 - Fix a bug: empty advanced accounting code is not defaulted to the general setting
 - Fix a bug: invalid style in accounting codes settings